REL-2859 Make ImageLoadingListener run inside a Monadic context #time 30m
